Fix security questions refering to text nodes.
This was broken by a82d1a8c, which tried to make page.text accept either
text nodes or element nodes, but instead made it not work properly for
text nodes (only element nodes). This broke all scripts that referred to
the text node in e.security.questions, which is essentially all of the
ones that have security questions.